requirements.txt
# Core
python-dotenv>=1.0.0

# LLM & Agents
ollama>=0.3.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0

# Browser automation
playwright>=1.49.0
browserbase>=1.0.0

# Web scraping
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Vector DB & RAG
chromadb>=0.5.0
pypdf>=5.0.0
python-docx>=1.0.0
openpyxl>=3.1.0

# UI
streamlit>=1.40.0
plotly>=5.24.0
pydeck>=0.9.0

# Database
# SQLite is built-in

# Utilities
pandas>=2.2.0
numpy>=1.26.0
geopy>=2.4.0
requests>=2.32.0
reportlab>=4.2.0
